+++ napari/_app_model/constants/__init__.py
