+++ napari/_app_model/constants/_commands.py
