+++ napari/_app_model/constants/_menus.py
