+++ napari/_app_model/context/_context.py
