+++ napari/_qt/qt_event_loop.py
@@ -194,7 +194,7 @@ def get_app(
         perf_config.patch_callables()
 
     if not _app_ref:  # running get_app for the first time
-        from napari import _app  # noqa   # TEMPORARY
+        from napari import _app_model  # noqa   # TEMPORARY
 
         # see docstring of `wait_for_workers_to_quit` for caveats on killing
         # workers at shutdown.
